CG_DrawNumField: Improve positioning precision

+++ code/cgame/cg_drawtools.c
@@ -548,7 +548,7 @@ Take x,y positions as if 640 x 480 and scales them to the proper resolution
 
 ==============
 */
-void CG_DrawNumField (int x, int y, int width, int value,int charWidth,int charHeight,int style,qboolean zeroFill)
+void CG_DrawNumField (float x, float y, int width, int value,int charWidth,int charHeight,int style,qboolean zeroFill)
 {
 	char	num[16], *ptr;
 	int		l;
